Fix getCurrentBranchRemote when not behind remote
Previous logic assumed the formatting

    mybranch...origin/mybranch [1 behind]

But, when the branches are on the same commit, the syntax is actually

    mybranch...origin/mybranch

This causes the parsing logic to fail, and not parse correctly, which
breaks the Git commands following.

Now, check for the presence of [ and skip the logic to cut the string
again if it isn't present.